+++ src/app/(auth)/sign-in/page.tsx
@@ -1,5 +1,5 @@
 import { SignInCard } from "@/features/auth/components/sign-in-card";
-import { getCurrent } from "@/features/auth/actions";
+import { getCurrent } from "@/features/auth/queries";
 import { redirect } from "next/navigation";
 
 const SignInPage = async () => {
