+++ src/app/(dashboard)/workspaces/[workspaceId]/page.tsx
@@ -1,4 +1,4 @@
-import { getCurrent } from "@/features/auth/actions";
+import { getCurrent } from "@/features/auth/queries";
 import { redirect } from "next/navigation";
 
 const WorkSpaceIdPage = async () => {
