+++ src/app/(standalone)/workspaces/[workspaceId]/settings/page.tsx
@@ -1,5 +1,5 @@
-import { getCurrent } from "@/features/auth/actions";
-import { getWorkspace } from "@/features/workspaces/actions";
+import { getCurrent } from "@/features/auth/queries";
+import { getWorkspace } from "@/features/workspaces/queries";
 import { EditWorkspaceForm } from "@/features/workspaces/components/edit-workspace-form";
 import { redirect } from "next/navigation";
