+++ src/features/auth/queries.ts
@@ -1,0 +1,10 @@
+import { createSessionClient } from "@/lib/appwrite";
+
+export const getCurrent = async () => {
+	try {
+		const { account } = await createSessionClient();
+		return await account.get();
+	} catch {
+		return null;
+	}
+};
