+++ src/features/workspaces/queries.ts
@@ -1,24 +1,12 @@
-import { cookies } from "next/headers";
-import { Account, Client, Databases, Query } from "node-appwrite";
-import { AUTH_COOKIE } from "../auth/constants";
+import { Query } from "node-appwrite";
 import { DATABASE_ID, MEMBERS_ID, WORKSPACES_ID } from "@/config";
 import { getMember } from "../members/utils";
 import { Workspace } from "./types";
+import { createSessionClient } from "@/lib/appwrite";
 
 export const getWorkspaces = async () => {
 	try {
-		const client = new Client()
-			.setEndpoint(process.env.NEXT_PUBLIC_APPWRITE_ENDPOINT!)
-			.setProject(process.env.NEXT_PUBLIC_APPWRITE_PROJECT!);
-		const session = await cookies().get(AUTH_COOKIE);
-
-		if (!session) {
-			return { documents: [], total: 0 };
-		}
-
-		client.setSession(session.value);
-		const databases = new Databases(client);
-		const account = new Account(client);
+		const { databases, account } = await createSessionClient();
 		const user = await account.get();
 		const members = await databases.listDocuments(DATABASE_ID, MEMBERS_ID, [
 			Query.equal("userId", user.$id),
@@ -46,18 +34,7 @@ interface GetWorkspaceProps {
 
 export const getWorkspace = async ({ workspaceId }: GetWorkspaceProps) => {
 	try {
-		const client = new Client()
-			.setEndpoint(process.env.NEXT_PUBLIC_APPWRITE_ENDPOINT!)
-			.setProject(process.env.NEXT_PUBLIC_APPWRITE_PROJECT!);
-		const session = await cookies().get(AUTH_COOKIE);
-
-		if (!session) {
-			return null;
-		}
-
-		client.setSession(session.value);
-		const databases = new Databases(client);
-		const account = new Account(client);
+		const { databases, account } = await createSessionClient();
 		const user = await account.get();
 		const member = await getMember({
 			databases,
