+++ src/lib/appwrite.ts
@@ -1,5 +1,7 @@
 import "server-only";
-import { Account, Client } from "node-appwrite";
+import { Account, Client, Databases } from "node-appwrite";
+import { cookies } from "next/headers";
+import { AUTH_COOKIE } from "@/features/auth/constants";
 
 export async function createAdminClient() {
 	const client = new Client()
@@ -9,6 +11,25 @@ export async function createAdminClient() {
 	return {
 		get account() {
 			return new Account(client);
+		},
+	};
+}
+
+export async function createSessionClient() {
+	const client = new Client()
+		.setEndpoint(process.env.NEXT_PUBLIC_APPWRITE_ENDPOINT!)
+		.setProject(process.env.NEXT_PUBLIC_APPWRITE_PROJECT!);
+	const session = await cookies().get(AUTH_COOKIE);
+	if (!session || !session.value) {
+		throw new Error("Unauthorized");
+	}
+	client.setSession(session.value);
+	return {
+		get account() {
+			return new Account(client);
+		},
+		get databases() {
+			return new Databases(client);
 		},
 	};
 }
